Function day06::part2

pub fn part2(input: &str) -> usize

Part Two

Your device’s communication system is correctly detecting packets, but still isn’t working. It looks like it also needs to look for messages.

A start-of-message marker is just like a start-of-packet marker, except it consists of 14 distinct characters rather than 4.

Here are the first positions of start-of-message markers for all of the above examples:

mjqjpqmgbljsphdztnvjfqwrcgsmlb: first marker after character 19

bvwbjplbgvbhsrlpgdmjqwftvncz: first marker after character 23

nppdvjthqldpwncqszvftbrmjlhg: first marker after character 23

nznrnfrfntjfmvfwmzdfjlvtqnbhcprsg: first marker after character 29

zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw: first marker after character 26

Example

How many characters need to be processed before the first start-of-message marker is detected?

assert_eq!(part2("mjqjpqmgbljsphdztnvjfqwrcgsmlb"), 19);
assert_eq!(part2("bvwbjplbgvbhsrlpgdmjqwftvncz"), 23);
assert_eq!(part2("nppdvjthqldpwncqszvftbrmjlhg"), 23);
assert_eq!(part2("nznrnfrfntjfmvfwmzdfjlvtqnbhcprsg"), 29);
assert_eq!(part2("zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw"), 26);

Panics

Panics if the input is less than 14 characters long.